Suppresion de jbump, amélioration éditeur, ajout traductions..

+++ core/src/main/kotlin/be.catvert.pc/utility/Point.kt
@@ -1,8 +1,9 @@
 package be.catvert.pc.utility
 
+
 /**
  * Classe de donnée représentant un point dans l'espace
  */
-data class Point(val x: Float = 0f, val y: Float = 0f) {
+data class Point(val x: Int = 0, val y: Int = 0) {
     override fun toString(): String = "{ $x ; $y }"
 }
